bug 7894 [SIGNED-OFF]: crude fix: convert results array into an arrayref
Signed-off-by: Magnus Enger <magnus@enger.priv.no>
This patch fixes the problem with the first record on the second
page, but there is still a problem with the next/previous links,
as described in comments 3 and 4 in the bug report. I'm signing
off since it's unclear wether that problem is directly linked to
the "first record on the second page" problem, and a partial fix
seems better than no fix at all.

Signed-off-by: Magnus Enger <magnus@enger.priv.no>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>