Bug 3413: Checkboxes for new subfields not saving correctly
The bug was in the numbering of the fields and subfields in the form.
The form elements for the 'new subfield' got the same numbered names and ids as
the form elements for the second subfield. (example: repeatable1)
The patch fixes the numbering to continue to the new subfield.

To test:
- Create some new subfields and check all checkbox fields save correctly
  (mandatory, repeatable, is url)
- Check that other subfields' checkboxes are not changed
- Check existing subfields and newly created subfields can be edited

Signed-off-by: mveron <veron@veron.ch>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>