Bug 7693 - clear screen button misaligned
This patch removes firefox specific lines in the css which were breaking the display

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

This fixes the layout problem in current versions of Firefox. I think
the CSS which is removed by this patch might be necessary for older
versions of Firefox, but if this turns out to be a problem we can
follow up.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>