Bug 5184 [REVISED] Upgrade jQuery to the latest version (OPAC)
Upgrade jQuery in the OPAC.

To test: Try any JavaScript-dependent interaction
you can think of in the OPAC and watch for errors.

Revision updates the one syntax I've found so far in testing
which requires an update. Where we used to do this to
uncheck a checkbox:

$(foo).attr("checked","");

We should now use:

$(foo).removeAttr("checked");

As far as I can tell the old way will not generate an error
but neither will it work.

Signed-off-by: Chris Nighswonger <cnighswonger@foundations.edu>