Bug 8014 - On the patron entry form hide "restricted until" field if "Restricted: No" is checked
This patch adds JavaScript to automatically hide the "Restricted
unil" form field on the patron entry form if "Restricted: No" is
selected. If "Restricted: Yes" is chosen the "Until:" field
is shown and focus is given to it.

The patch also adds a note, "(optional)" after the "Until:" field
to make it clearer that a date is not required.

To test, open a patron record and try moving the selection
between "Restricted: No" and "Restricted: Yes." Confirm that
the "Until:" field is shown and hidden as expected. Confirm
that the restricted flag is correctly set or unset upon save,
including setting the flag with or without a date.

Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>