Bug 8035 - bibs with comments show an error in opac
If a record has any approved comments, viewing the record
details in the opac will error out with the following
message:

Undefined subroutine &main::format_date called at
/home/load02/kohaclone/opac/opac-detail.pl line 654.

This is caused by a use of C4::Dates::format_date
without adding a use call to the head of the file.

Removing the call is the only change necessary. The
dates are already being filtered by $KohaDates.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>