ug 8078 [REVISED] my summary showing contents of incorrect tab
This patch moves the holdings tab content to the correct
position sequentially with regard to the order of the
tabs themselves. My guess is that the jQuery upgrade has
something to do with this change but I'm not sure why
(probably something in our aging tabs plugin which works
differently with the new jQuery version).

Making the sequence of the tab contents match the sequence
of the tabs seems to fix the problem.

Revised to correct the error in placement of the tags markup
in the staff client.

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>