Bug 5888: opac-detail subject/author links improvements
When the user clicks on a subject or an author, the list of subjects
(authors) is displayed and the user can choose more than 1 subject
(author) to search on.
Also added a link to opac-authoritiesdetail.pl when possible.

Only affects 'Normal view'. Does not affect XSLT display.

Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>
Split out of the original patch for bug 5888. The links to the
authority records are added by bug 8210.

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>