Bug 8160 - Link to cataloging appears for users without cataloging permission
This patch wraps cataloging links on the staff client home
page and in the header menu with checks for edit_catalogue
OR edit_items permission.

To test, access the staff client as a user with various
permissions. The "cataloging" link on the main page and
in the header menu should only appear to a user with
edit_catalogue permission, edit_items permission, or both.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>