bug 8323 fixing year sorting of a list
Signed-off-by: Marc Veron <veron@veron.ch>

Works as expected.

Caveat: If you edit the list in Opac, change the sort and then klick 'Save',
the List always is displayed in title order. This is du to the fact that
save calls opac-shelves.pl?viewshelf=4 without the sort parameter (defaults
to 'title').
If you go back to lists, choose the list from List name, the param is
correct and the list displays in expected order.

It took quite a time to track  that down, I first thought that opac-shelves.pl
was broken.

However, current bug 8323 is not affected, signing off.

Marc

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>