Bug 8301 - Display additional system information (preferences) on About page to minimize support cases
To minimize support questions regarding common mistakes, the About
page should have an additional tab that displays warnings if depractated
and/or mutual excluding system preferences are used.

To test patch use values as displayed on screenshot attached.

New: Shows error message in Tab Server Information if Zebra server is not started
(...and changes added...)

Suggestions for other values to add are welcome.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>