Bug 8207: Allow see also fields in auths to link to thesauri
Right now if you were to link a field in an authority to a thesaurus, it
would not work properly. The authority type frameworks require some
adjustments to allow see also headings to be linked to thesauri (such as
adding subfield $9).

This patch adds subfield $9 and thesaurus links to existing installs
for those 5xx fields (under MARC21 and NORMARC only) which do not yet
have any subfield $9 or thesauri configured, as well as adding that to
new English language installs. There will be a follow-up patch adding
the subfields to new non-English language install files later, once
it is clear that the changes proposed and demonstrated in the English
files are accepable to the community.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Tested frameworks install cleanly on a new installation.

Tested database update and made sure, that configuration and subfields
of a field were not changed if
- a $9 subfield already existed for a 5xx field
- a $9 subfield and a link to another thesaurus in $a existed
- no $9 but a link to another thesaurus in $a existed

This seems safe to me.