Bug 8398 - Missing image file in staff global css
This patch fixes a regression introduced by the fix for
Bug 6855 (bullet point showing in wrong place on error message)
http://bugs.koha-community.org/bugzilla3/show_bug.cgi?id=6855

The patch removes the re-introduced reference to a
background image which has since been removed. Testers should
check that this patch does not re-introduce Bug 6855.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
I checked the styling of the error messages in Chromium and Firefox.
The logs are no longer complaining about the missing image.
Looks good to me.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>