Bug 8408 - Odd OPAC message when OPAC Renewals are disabled and OPACFineNoRenewals is set to blank
If OpacRenewalAllowed is set to "Don't allow" and OPACFineNoRenewals is set to blank ( i.e. disabled ).
A user who owes any fines, when logged in to the OPAC will receive the following message:

    Please note: You have more than 0.00 in fines.

This is a rather odd message, and is not very sensible. The message should either be modified or removed altogether.

This patch modifies the message to the following:

    Please note: You currently owe $fines in fines

where $fines in the formatted amount the borrower currently owes
in fines.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>