Bug 8206: make authority search indexes consistent
This commit makes the search indexes used for search authorities in the
staff client and OPAC more consistent by using the Mainentry,
Match, and Any indexes for both, and adjusting the labels so that they
correctly describe what the indexes do.

Note that the Match index was chosen instead of the Match-heading index
because Match-heading has the additional attribute of being for phrase
searches and exact matches, whereas Match is used for keyword matching,
and is more applicable with GRS-1 indexed records (thanks to the
well-known limitations of GRS-1 and exact matching of headings).

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>
No problem detected in UNIMARC