Bug 3374 - Display patron attributes in the same format as other patron data
This patch revises the interface for displaying and editing
patron attributes so that it is consistent with other
similar interfaces.

Display of patron attributes should match the display of other
kinds of patron data, and codes should be omitted in favor of
descriptions.

Entry of patron attributes should match other patron entry
forms and should also omit codes in favor of descriptions.

To test, view and edit patron attributes. Special functions
like "Clear" and "New" should function as before.

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
Tested by viewing patrons with attributes, editing those attributes
and editing the entire patron record. In all places the editing/adding
works and the display is improved.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>