Bug 6617: Improve MARC21 enhanced contents display
The display of "Enhanced" 505 (contents) fields in the MARC21
XSLT is very poor, resulting in large, unreadable blocks of text,
and -- for some series -- so much unnecessary duplication in the
Notes tab that important information is entirely obscured. This
patch reformats "enhanced" contents fields (MARC21 505 fields with
$t and $r) so as to be more readable by breaking up entries on
separate lines and making titles bold (to make them stand out more).

This patch does not address duplication of information in the Notes
tab, per discussion on the #koha IRC channel.

To test:
1) View record with enhanced 505 field before applying patch. Observe
   it is very unfriendly.
2) Apply patch.
3) View same record, note that display is much improved.

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>

This first patch works as advertised. It formats the table of
contents on the opac detail display in a cleaner format.

I recommend that future follow ups fix the staff client to match
and update the title notes tabs on the opac and staff client
to also match the cleaner format for table of contents.