Bug 8617 - Problem with display of comments waiting for moderation in OPAC
This patch addresses on aspect of Bug 8617: The behavior where
live previews of edited comments are displaying on opac-detail
as a duplicate comment. We must pass the review id back to the
script when submitting so that the check for it will evaluate
as true when it should.

This patch also corrects some JavaScript for translatability.

To test, submit a comment on a title and approve it. Return to
the title and edit the comment. When you submit your revised
comment the comment pop-up window should close and your revised
comment should appear on opac-detail as "pending approval."

Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>