Bug 8656 Calendar needs to load both exception and nonexception dates
single_holidays and exception_holidays were both being selected
with the same parameter with the result that single holidays were not
being loaded.

Have removed the retrieval of the unused parameters title and
description.

In effect the distinction is not meaningful for these routines and
the two structures could become one. This patch just fixes the
current bug

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>