Bug 8664 - Add floating toolbar with save button to system preferences editor
This patch adds a copy of the "save" button in a toolbar at
the top of the system preferences editor. The toolbar becomes
fixed at the top of the page when you scroll down, exactly like
the toolbar on addbiblio.pl.

Because multiple save buttons might appear when the system prefs
page displays search results, the toolbar does not appear in
this case.

To test, open the system preferences editor and change a setting.
Try saving both with the button in the toolbar and the button at
the top of the page. Test scrolling to confirm that the toolbar
"sticks" to the top of the page when you scroll down.

Signed-off-by: Mirko Tietgen <mirko@abunchofthings.net>
Passed-QA-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>